Add Pinboard update command.
The contents of this function were copied from
[jacobian/pinboard-to-sqlite](https://github.com/jacobian/pinboard-to-sqlite/blob/master/pinboard_to_sqlite/cli.py)
because I couldn't make a direct call to the posts() function work as a
module.  Problably because of conflicting Click directives.  There is
some refactoring needed here--particularly to remove the dependency on
sqlite_utils and add error handling for the Pinboard API call.