Recreate links to new sources on daily page
The thought being that I would review the automatically generated
summary and adjust the wiki links/tags before removing the internal wiki
link markup from the page.  Long term, the extra link between the daily
page and the source is semantic clutter, but it is useful as a reminder
and shortcut to check "Yesterday's readings" on the daily journal page.